python/requirements.txt
PyQt5==5.12.0
pytest==3.4.1
python-can==4.1.0
